New FeaturedBlogPosts component and updated featured blog posts list
